Adding StepUtils and refactoring code
