+++ src/test/java/features/ViaCep.feature
@@ -1,7 +1,13 @@
-Feature: Test consult the via CEP API
+Feature: Test consult the via CEP API - Scenario Outline
 
   @viaCep
-  Scenario: Test Get viaCEP 2
-    Given User calls "viaCEP2" with "GET" http request
+  Scenario Outline: Test Get viaCEP 2
+    Given an address code <cep>
+    When User calls "viaCEP1" with "GET" http request
     Then Status code is 200
     And Validate the schema "viaCEP"
+    Examples:
+      | cep        |
+      | "12248310" |
+      | "12248330" |
+      | "12248320" |
