+++ src/test/java/resources/APIResources.java
@@ -4,7 +4,7 @@ public enum APIResources {
     AddPlaceAPI("/maps/api/place/add/json"),
     DeletePlaceAPI("/maps/api/place/delete/json"),
     viaCEP("/ws/12246870/json"),
-    viaCEP1("/ws/12246870/json"),
+    viaCEP1("/ws/{cep}/json"),
     viaCEP2("/ws/12246873/json");
     private final String resource;
